requirements.txt
dnspython>=2.4.0
python-whois>=0.8.0
requests>=2.31.0
rich>=13.7.0
shodan>=1.31.0
